pyproject.toml
[tool.poetry]
name = "asig"
version = "0.1.0"
description = ""
authors = ["Daniel Cuznetov <daniel.cuznetov04@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.6.0"
zeep = "^4.3.1"
gunicorn = "^23.0.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
pymupdf = "^1.25.1"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py37"
exclude = [
    "*/migrations/*"
]

lint.select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "ASYNC", # flake8-async
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "BLE", # flake8-blind-except
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "FIX", # flake8 - fix me
    "N", # pep8-naming
    "UP", # pyupgrade
]

lint.ignore = [
    "N803", "N806"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
